Reorganize headers into `sass` subfolder
