+++ Readme.md
@@ -93,4 +93,3 @@ Our MIT license is designed to be as simple, and liberal as possible.
 sass2scss was originally written by [Marcel Greter](@mgreter)
 and he happily agreed to have it merged into the project.
 
-[sass_interface.h]: sass_interface.h
