+++ src/GNUmakefile.am
@@ -44,7 +44,8 @@ endif
 
 include_HEADERS = $(top_srcdir)/include/sass.h \
                   $(top_srcdir)/include/sass2scss.h \
-                  $(top_srcdir)/include/sass_values.h \
-                  $(top_srcdir)/include/sass_version.h \
-                  $(top_srcdir)/include/sass_context.h \
-                  $(top_srcdir)/include/sass_functions.h
+                  $(top_srcdir)/include/sass/base.h \
+                  $(top_srcdir)/include/sass/values.h \
+                  $(top_srcdir)/include/sass/version.h \
+                  $(top_srcdir)/include/sass/context.h \
+                  $(top_srcdir)/include/sass/functions.h
