+++ src/eval.cpp
@@ -14,7 +14,7 @@
 #include "inspect.hpp"
 #include "environment.hpp"
 #include "position.hpp"
-#include "sass_values.h"
+#include "sass/values.h"
 #include "to_value.hpp"
 #include "to_c.hpp"
 #include "context.hpp"
