+++ src/to_value.cpp
@@ -1,5 +1,4 @@
 #include "ast.hpp"
-#include "sass_values.h"
 #include "to_value.hpp"
 #include "to_string.hpp"
